requirements.txt
Django==1.6.2
PyYAML==3.11
South==0.8.4
argh==0.24.1
django-haystack==2.1.0
django-static-precompiler==0.4
pathtools==0.1.2
pyelasticsearch==0.6.1
requests==2.2.1
simplejson==3.4.0
six==1.6.1
wsgiref==0.1.2
